--- server/utils.js
@@ -1,6 +1,0 @@
-const jwt = require('jsonwebtoken');
-
-exports.generateAuthToken = (payload, secret, expiresIn) =>
-  jwt.sign(payload, secret, {
-    expiresIn
-  });
