+++ src/typeDefs.ts
@@ -1,4 +1,4 @@
-module.exports = `
+export default `
   type AuthPayload {
     token: String!
   }
