+++ src/utils.ts
@@ -1,0 +1,6 @@
+import * as jwt from 'jsonwebtoken';
+
+export const generateAuthToken = (payload, secret, expiresIn) =>
+  jwt.sign(payload, secret, {
+    expiresIn
+  });
