+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "moduleResolution": "node",
+    "module": "commonjs",
+    "sourceMap": true,
+    "noUnusedLocals": true,
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "lib": ["esnext.asynciterable", "es2016", "dom"]
+  },
+  "exclude": ["node_modules", "dist", "example"]
+}
